Telegram_bot Voice Notebook with Whisper/requirements.txt
anyio==4.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
exceptiongroup==1.2.1
filelock==3.13.1
fsspec==2024.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.3
llvmlite==0.42.0
MarkupSafe==2.1.5
mkl==2021.4.0
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.2.1
numba==0.59.1
numpy==1.26.3
openai-whisper==20231117
pillow==10.2.0
python-telegram-bot==21.2
regex==2024.5.15
requests==2.32.3
sniffio==1.3.1
sympy==1.12
tbb==2021.11.0
tiktoken==0.7.0
torch==2.3.0+cu121
torchaudio==2.3.0+cu121
torchvision==0.18.0+cu121
tqdm==4.66.4
typing_extensions==4.9.0
urllib3==2.2.1
